% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmedians.direction.R
\name{Cmedians.direction}
\alias{Cmedians.direction}
\title{Tests for a Common Median Direction between groups}
\usage{
Cmedians.direction(clustr, traj, id_field = TRUE, Nsample = 999)
}
\arguments{
\item{clustr}{[vector (charater)] A vector of cluster membership (labels). For instance, the result extracted from the \code{\link{akmedoids.clust}} function.}

\item{traj}{[matrix (numeric)]: corresponding longitudinal data used to generate \code{clustr} (with rows corresponding to each label of \code{clustr}). For example, the first label of \code{clustr} is the group label of the first row of \code{traj} matrix, and so on.}

\item{id_field}{[numeric or character] Whether the first column of the \code{traj} is a unique (\code{id}) field. Default: \code{FALSE}. If \code{TRUE} the function recognises the second column as the first time step.}

\item{Nsample}{[an integer (numeric)] Number of random permutation of the original combined sample}
}
\value{
A matrix showing the pvalues indicating whether the medians direction of any two groups are drawn from the same distribution.
}
\description{
This function tests the null hypothesis of a common median direction between pairs of groups resulting from the \code{akmedoids.clust} function. The aim is to provide statistical significance values that could be used to further categorise group into three large classes, denoted as the 'Decreasing', 'Stable' & 'Increasing' classes.
}
\details{
This function implements the Fisher's Nonparametric Test (Fisher, 1993) to determine whether g independent samples were drawn from distributions with a common median direction.
}
\examples{
print(traj)
traj <- dataImputation(traj, id_field = TRUE, method = 1, replace_with = 1,
fill_zeros = FALSE)
print(traj)
traj <- props(traj, id_field = TRUE)
clustr <- akmedoids.clust(traj, id_field = TRUE, method = "linear", k = c(3,6))
clustr <- as.vector(clustr$optimSolution)
p.values <- Cmedians.direction(clustr, traj, id_field=TRUE, Nsample=999)
}
\references{
\code{1}. Fisher N. I. (1993). Statistical Analysis of Circular Data. Cambridge University Press, Cambridge.

\code{2}. Pewsey A. et al. (2013). Circular Statistics in R. Oxford University Press (1st Edition).
}
