% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinusPiPi.R
\name{MinusPiPi}
\alias{MinusPiPi}
\title{Tests}
\usage{
MinusPiPi(sample)
}
\arguments{
\item{sample}{[matrix (numeric)]:}
}
\value{
A matrix
}
\description{
This function
}
\details{
This function
}
\references{
\code{1}. Fisher N. I. (1993). Statistical Analysis of Circular Data. Cambridge University Press, Cambridge.

\code{2}. Pewsey A. et al. (2013). Circular Statistics in R. Oxford University Press (1st Edition).
}
