% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgRandTest.R
\name{PgRandTest}
\alias{PgRandTest}
\title{Tests}
\usage{
PgRandTest(cdat, ndat, g, Nsample)
}
\arguments{
\item{cdat}{[matrix (numeric)]:}

\item{ndat}{[numeric or character]}

\item{g}{[an integer (numeric)]}

\item{Nsample}{[an integer (numeric)] Number}
}
\value{
A matrix
}
\description{
This function
}
\details{
This function
}
\references{
\code{1}. Fisher N. I. (1993). Statistical Analysis of Circular Data. Cambridge University Press, Cambridge.

\code{2}. Pewsey A. et al. (2013). Circular Statistics in R. Oxford University Press (1st Edition).
}
