% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PgVal.R
\name{PgVal}
\alias{PgVal}
\title{Tests}
\usage{
PgVal(cdat, ndat, g)
}
\arguments{
\item{cdat}{[matrix (numeric)]:}

\item{ndat}{[numeric or character]}

\item{g}{[an integer (numeric)]}
}
\value{
A matrix
}
\description{
This function
}
\details{
This function
}
\references{
\code{1}. Fisher N. I. (1993). Statistical Analysis of Circular Data. Cambridge University Press, Cambridge.

\code{2}. Pewsey A. et al. (2013). Circular Statistics in R. Oxford University Press (1st Edition).
}
