% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope.diff.R
\name{slope.diff}
\alias{slope.diff}
\title{To calculate the difference between the angles of mean slopes of two trajectory objects}
\usage{
slope.diff(traj1, traj2, id_field = TRUE)
}
\arguments{
\item{traj1}{[matrix (numeric)]: corresp}

\item{traj2}{[matrix (numeric)]: corresp}

\item{id_field}{[ss] All must have id_field.}
}
\value{
A matrix showing the pvalues indicating whether the medians direction of any two groups are drawn from the same distribution.
}
\description{
This function extracts the mean slope a set of trajectories
}
\details{
This function implements the Fisher's Nonparametric Test (Fisher, 1993) to determine whether g independent samples were drawn from distributions with a common median direction.
}
