% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatsonU2TestRand.R
\name{WatsonU2TestRand}
\alias{WatsonU2TestRand}
\title{Tests for a Common Median Direction between groups}
\usage{
WatsonU2TestRand(clustr, traj, id_field = TRUE, grp = c(1, 2),
  Nsample = 999)
}
\arguments{
\item{clustr}{[matrix (numeric)]: corresponding longitudinal data used to generate \code{clustr} (with rows corresponding to each label of \code{clustr}). For example, the first label of \code{clustr} is the group label of the first row of \code{traj} matrix, and so on.}

\item{traj}{[matrix (numeric)]: corresponding longitudinal data used to generate \code{clustr} (with rows corresponding to each label of \code{clustr}). For example, the first label of \code{clustr} is the group label of the first row of \code{traj} matrix, and so on.}

\item{id_field}{[ss]}

\item{grp}{[sss]}

\item{Nsample}{[an integer (numeric)] Number of random permutation of the original combined sample}
}
\value{
A matrix showing the pvalues indicating whether the medians direction of any two groups are drawn from the same distribution.
}
\description{
This function tests the null hypothesis of a common median direction between pairs of groups resulting from the \code{akmedoids.clust} function. The aim is to provide statistical significance values that could be used to further categorise group into three large classes, denoted as the 'Decreasing', 'Stable' & 'Increasing' classes.
}
\details{
This function implements the Fisher's Nonparametric Test (Fisher, 1993) to determine whether g independent samples were drawn from distributions with a common median direction.
}
\examples{
print(traj)
traj <- dataImputation(traj, id_field = TRUE, method = 1, replace_with = 1,
fill_zeros = FALSE)
print(traj)
traj <- props(traj, id_field = TRUE)
clustr <- akmedoids.clust(traj, id_field = TRUE, method = "linear", k = c(3,6))
clustr <- as.vector(clustr$optimSolution)
p.values <- WatsonU2TestRand(clustr, traj, id_field=TRUE, grp=c(1:2), Nsample=999)
}
\references{
\code{1}. Fisher N. I. (1993). Statistical Analysis of Circular Data. Cambridge University Press, Cambridge.

\code{2}. Pewsey A. et al. (2013). Circular Statistics in R. Oxford University Press (1st Edition).
}
